+++ R/000.DEPRECATION.R
@@ -6,9 +6,9 @@ isUseNamesNADefunct <- function() {
 
 deprecatedUseNamesNA <- function() {
   if (isUseNamesNADefunct()) {
-    .Defunct(msg = "useNames = NA is defunct. Instead, specify either useNames = TRUE or useNames = FALSE.", package = .packageName)
+    .Defunct(msg = sprintf("[%s (>= 1.2.0)] useNames = NA is defunct. Instead, specify either useNames = TRUE or useNames = FALSE. See also ?matrixStats::matrixStats.options", .packageName), package = .packageName)
   } else {
-    .Deprecated(msg = "useNames = NA is deprecated. Instead, specify either useNames = TRUE or useNames = FALSE.", package = .packageName)
+    .Deprecated(msg = sprintf("[%s (>= 1.2.0)] useNames = NA is deprecated. Instead, specify either useNames = TRUE or useNames = FALSE. See also ?matrixStats::matrixStats.options", .packageName), package = .packageName)
   }
 }
 
@@ -16,44 +16,42 @@ deprecatedUseNamesNA <- function() {
 defunctShouldBeMatrixOrDim <- function(x) {
   x_class <- sQuote(class(x)[1])
   x_name <- sQuote(as.character(substitute(x)))
-  .Defunct(msg = sprintf("Argument %s is of class %s, but should be a matrix or 'dim.' should specify one. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", x_name, x_class, x_class))  #nolint
+  .Defunct(msg = sprintf("[%s] Argument %s is of class %s, but should be a matrix or 'dim.' should specify one. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", .packageName, x_name, x_class, x_class))  #nolint
 }
 
 defunctShouldBeMatrix <- function(x) {
   x_class <- sQuote(class(x)[1])
   x_name <- sQuote(as.character(substitute(x)))
-  .Defunct(msg = sprintf("Argument %s is of class %s, but should be a matrix. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", x_name, x_class, x_class))  #nolint
+  .Defunct(msg = sprintf("[%s] Argument %s is of class %s, but should be a matrix. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", .packageName, x_name, x_class, x_class))  #nolint
 }
 
 defunctShouldBeMatrixOrVector <- function(x) {
   x_class <- sQuote(class(x)[1])
   x_name <- sQuote(as.character(substitute(x)))
-  .Defunct(msg = sprintf("Argument %s is of class %s, but should be a matrix or a vector. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", x_name, x_class, x_class))  #nolint
+  .Defunct(msg = sprintf("[%s] Argument %s is of class %s, but should be a matrix or a vector. The use of a %s is not supported, the correctness of the result is not guaranteed. Please update your code accordingly.", .packageName, x_name, x_class, x_class))  #nolint
 }
 
 validateScalarCenter <- function(center, n, dimname) {
-  onScalar <- getOption("matrixStats.center.onScalar", NULL)
+  onScalar <- getOption("matrixStats.center.onScalar", "defunct")
+  if (identical(onScalar, "ignore")) return()
   
-  ## Nothing to do?
-  if (is.null(onScalar)) return()
-
-  fcn <- switch(onScalar, deprecated = .Deprecated, defunct = .Defunct, NULL)
+  action <- switch(onScalar,
+    deprecated = .Deprecated,
+       defunct = .Defunct,
+    function(...) NULL
+  )
   
-  ## Nothing to do?
-  if (is.null(fcn)) return()
-
-  msg <- sprintf("Argument '%s' should be of the same length as number of %s of '%s'. Use of a scalar value is deprecated: %s != %s", "center", dimname, "x", length(center), n)
-  fcn(msg = msg, package = .packageName)
+  msg <- sprintf("[%s (>= 0.58.0)] Argument '%s' should be of the same length as number of %s of '%s'. Use of a scalar value is %s: %s != %s (See also ?matrixStats::matrixStats.options)", .packageName, "center", dimname, "x", onScalar, length(center), n)
+  action(msg = msg, package = .packageName)
 }
 
 
 validateVarsCenterFormula <- local({
-  .curr <- 1
-  .next <- 1
+  countdown <- NA_integer_
   always <- structure(TRUE, when = "each time this function is called")
   
   function() {
-    freq <- getOption("matrixStats.vars.formula.freq", NULL)
+    freq <- getOption("matrixStats.vars.formula.freq", 50L)
     
     ## Nothing to do?
     if (is.null(freq)) return(FALSE)
@@ -67,83 +65,81 @@ validateVarsCenterFormula <- local({
     ## each time?
     if (freq == 1) return(always)
 
-    ## once in a while?
-    .curr <<- .curr + 1
-    .next <<- freq  ## update .next according to R option
-
-    ## Skip or not?
-    if (.curr <= .next) return(FALSE)
-    .curr <<- 1 ## reset
+    ## Not first incident?
+    if (!is.na(countdown)) {
+      countdown <<- countdown - 1L
+      if (countdown > 0L) return(FALSE)
+    }
+    countdown <<- freq ## reset
     structure(TRUE, when = sprintf("every %g call to this function", freq))
   }
 })
 
 
 centerOnUse <- function(fcnname, calls = sys.calls(), msg = NULL) {
-  value <- getOption("matrixStats.center.onUse")
-  if (is.null(value) || identical(value, "ignore")) return()
+  value <- getOption("matrixStats.center.onUse", "ignore")
+  if (identical(value, "ignore")) return()
   
   value <- match.arg(value, c("deprecated", "defunct"))
-  fcn <- switch(value, deprecated = .Deprecated, defunct = .Defunct)
+  action <- switch(value,
+    deprecated = .Deprecated,
+       defunct = .Defunct,
+    function(...) NULL
+  )
 
   if (is.null(msg)) {
-    msg <- sprintf("Argument '%s' of %s::%s() is %s: %s", "center",
-                   .packageName, fcnname, value, deparse(calls[[1]])[1])
+    msg <- sprintf("[%s] Argument '%s' of %s::%s() is %s: %s (See also ?matrixStats::matrixStats.options)",
+                   .packageName, "center", .packageName, fcnname,
+                   value, deparse(calls[[1]])[1])
   }
+  action(msg = msg, package = .packageName)
+}
+
+
+
+validateTiesMethodMissing <- local({
+  countdown <- NA_integer_
+  always <- structure(TRUE, when = "each time this function is called")
   
-  fcn(msg = msg, package = .packageName)
-}
+  function() {
+    freq <- getOption("matrixStats.ties.method.freq", 50L)
+    
+    ## Nothing to do?
+    if (is.null(freq)) return(FALSE)
 
+    ## never?
+    if (freq <= 0) return(FALSE)
 
-onLoadSetVarsFormulaFreq <- function() {
-  ## Option is already set?
-  if (!is.null(getOption("matrixStats.vars.formula.freq", NULL))) return()
+    ## always?
+    if (is.infinite(freq)) return(always)
 
-  ## Is environment variable set?
-  value <- Sys.getenv("R_MATRIXSTATS_VARS_FORMULA_FREQ", "50")
-  if (nzchar(value)) {
-    value <- as.numeric(value)
-  } else {
-    value <- NULL
+    ## each time?
+    if (freq == 1L) return(always)
+
+    ## Not first incident?
+    if (!is.na(countdown)) {
+      countdown <<- countdown - 1L
+      if (countdown > 0L) return(FALSE)
+    }
+    countdown <<- freq ## reset
+    structure(TRUE, when = sprintf("every %g call to this function", freq))
   }
+})
 
-  options(matrixStats.vars.formula.freq = value)
-}
+tiesMethodMissing <- local({
+  function() {
+    action <- getOption("matrixStats.ties.method.missing", if (getRversion() >= "4.4.0") "deprecated" else "ignore")
+    if (action == "ignore") return()
 
-
-onLoadSetVarsFormulaOnMistake <- function() {
-  ## Option is already set?
-  if (!is.null(getOption("matrixStats.vars.formula.onMistake", NULL))) return()
-
-  ## Is environment variable set?
-  value <- Sys.getenv("R_MATRIXSTATS_VARS_FORMULA_MISTAKE", "defunct")
-  if (is.na(value)) return()
-
-  value <- match.arg(value, c("deprecated", "defunct"))
-  options(matrixStats.vars.formula.onMistake = value)
-}
-
-
-onLoadSetCenterOnUse <- function() {
-  ## Option is already set?
-  if (!is.null(getOption("matrixStats.center.onUse", NULL))) return()
-
-  ## Is environment variable set?
-  value <- Sys.getenv("R_MATRIXSTATS_CENTER_ONUSE", NA_character_)
-  if (is.na(value)) return()
-
-  value <- match.arg(value, c("ignore", "deprecated", "defunct"))
-  options(matrixStats.center.onUse = value)
-}
-
-onLoadSetCenterOnScalar <- function() {
-  ## Option is already set?
-  if (!is.null(getOption("matrixStats.center.onScalar", NULL))) return()
-
-  ## Is environment variable set?
-  value <- Sys.getenv("R_MATRIXSTATS_CENTER_ONSCALAR", "deprecated")
-  if (is.na(value)) return()
-
-  value <- match.arg(value, c("ignore", "deprecated", "defunct"))
-  options(matrixStats.center.onScalar = value)
-}
+    ## How often should we check?
+    if (!validateTiesMethodMissing()) return()
+    
+    action <- switch(action,
+      deprecated = .Deprecated,
+      defunct    = .Defunct,
+      function(...) NULL
+    )
+    msg <- sprintf("[%s] Please explicitly specify argument 'ties.method' when calling colRanks() and rowRanks() of %s. This is because the current default ties.method=\"max\" will eventually be updated to ties.method=\"average\" in order to align with the default of base::rank(). See also ?matrixStats::matrixStats.options", .packageName, .packageName)
+    action(msg = msg, package = .packageName)
+  }
+})